Make initial app - Day 2 session -- Forms: --- General Info: allows adding/editing fields directly --- EXP / EDU: allows adding items, editing items, deleting items. The forms are hidden and get displayed only base on user action from view part.
-- Views:
--- General Info: Contains only the form
--- EXP / EDU: Each contains form (initially hidden) and displays
any existing items. When clicking to add or edit, the form gets
displayed. When canceling/saving/editing, the action gets performed
and the form is hidden again.

-- Started with state in lower component, then lifted state up,
then just used context (CVContext.jsx). The app is wrapped in
CVProvider so that all components can interact with the context
they need.

-- Preview: Display all current info gathered from global state

-- Ignore prop validation warnings from ESLint for now